+++ Core/JsonContent.cs
@@ -429,7 +429,7 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public JsonContent Put(string name, IData v, int proj = 0x00ff)
+        public JsonContent Put(string name, IData v, short proj = 0x00ff)
         {
             if (counts[level]++ > 0) Add(',');
             if (name != null)
@@ -462,7 +462,7 @@ namespace Greatbone.Core
             return this;
         }
 
-        public JsonContent Put<D>(string name, D[] v, int proj = 0x00ff) where D : IData
+        public JsonContent Put<D>(string name, D[] v, short proj = 0x00ff) where D : IData
         {
             if (counts[level]++ > 0) Add(',');
             if (name != null)
